Move code from continuations.c to trampoline.c
